Cleanup Python extension
