+++ python/python/Ice/BatchRequestInterceptor.py
@@ -1,7 +1,4 @@
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) ZeroC, Inc. All rights reserved.
-#
 
 class BatchRequestInterceptor(object):
     """Base class for batch request interceptor. A subclass must
