+++ python/python/Ice/EnumBase.py
@@ -1,7 +1,4 @@
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) ZeroC, Inc. All rights reserved.
-#
 
 class EnumBase(object):
     def __init__(self, _n, _v):
