+++ python/python/Ice/Exception.py
@@ -1,7 +1,4 @@
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) ZeroC, Inc. All rights reserved.
-#
 
 class Exception(Exception):  # Derives from built-in base 'Exception' class.
     """The base class for all Ice exceptions."""
