--- python/python/Ice/FacetMap.py
@@ -1,19 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright (c) ZeroC, Inc. All rights reserved.
-#
-
-import Ice
-import IcePy
-
-# Start of module Ice
-_M_Ice = Ice.openModule("Ice")
-__name__ = "Ice"
-
-# TODO can we remove this? It's not used in the generated code.
-if "_t_FacetMap" not in _M_Ice.__dict__:
-    _M_Ice._t_FacetMap = IcePy.defineDictionary(
-        "::Ice::FacetMap", (), IcePy._t_string, IcePy._t_Value
-    )
-
-# End of module Ice
