+++ python/python/Ice/Future.py
@@ -1,8 +1,4 @@
-
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) ZeroC, Inc. All rights reserved.
-#
 
 import threading
 import time
