+++ python/python/Ice/InvocationFuture.py
@@ -1,7 +1,4 @@
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) ZeroC, Inc. All rights reserved.
-#
 
 import logging
 import traceback
