+++ python/python/Ice/LocalException.py
@@ -1,7 +1,4 @@
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) ZeroC, Inc. All rights reserved.
-#
 
 class LocalException(Exception):
     """The base class for all Ice run-time exceptions."""
