+++ python/python/Ice/LocalExceptions.py
@@ -1,7 +1,4 @@
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) ZeroC, Inc. All rights reserved.
-#
 
 from .LocalException import LocalException
