+++ python/python/Ice/ModuleUtil.py
@@ -1,6 +1,4 @@
-#
 # Copyright (c) ZeroC, Inc. All rights reserved.
-#
 
 import sys
 import types
