+++ python/python/Ice/PropertiesAdminUpdateCallback.py
@@ -1,7 +1,4 @@
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) ZeroC, Inc. All rights reserved.
-#
 
 class PropertiesAdminUpdateCallback(object):
     """Callback class to get notifications of property updates passed
